requirements.txt
requests
python-dotenv
pypdf2
python-docx
pdfplumber
python-telegram-bot==13.7
openai